[BACKLOG-21154] Use whiteboard pattern on Pentaho Web Packages
