+++ pentaho-webpackage/core/api/src/main/java/org/pentaho/webpackage/core/PentahoWebPackage.java
@@ -21,8 +21,4 @@ public interface PentahoWebPackage {
   String getVersion();
 
   String getResourceRootPath();
-
-  void init();
-
-  void destroy();
 }
