+++ pentaho-webpackage/core/impl/src/test/resources/invalid-package.json
