+++ pentaho-webpackage/core/impl/src/test/resources/pentaho-webpackage-1a-package.json
