+++ pentaho-webpackage/core/impl/src/test/resources/pentaho-webpackage-1b-package.json
