+++ pentaho-webpackage/core/impl/src/test/resources/pentaho-webpackage-2c-package.json
