+++ pentaho-webpackage/core/impl/src/test/resources/root-package.json
