--- pentaho-webpackage/extenders/pom.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-         xmlns="http://maven.apache.org/POM/4.0.0">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.pentaho.webpackage</groupId>
-    <artifactId>pentaho-webpackage-modules</artifactId>
-    <version>8.2.0.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>extenders-modules</artifactId>
-  <version>8.2.0.0-SNAPSHOT</version>
-
-  <packaging>pom</packaging>
-
-  <modules>
-    <module>http-extender</module>
-  </modules>
-</project>
