+++ pentaho-webpackage/listeners/http-listener/src/test/resources/invalid-package.json
@@ -1,0 +1,1 @@
+invalid content
