+++ pentaho-webpackage/listeners/http-listener/src/test/resources/pentaho-webpackage-1b-package.json
@@ -1,0 +1,4 @@
+{
+  "name": "package-name-1b",
+  "version": "1.1"
+}
