+++ pentaho-webpackage/listeners/http-listener/src/test/resources/pentaho-webpackage-1c-package.json
@@ -1,0 +1,4 @@
+{
+  "name": "package-name-1c",
+  "version": "1.2"
+}
