+++ pentaho-webpackage/listeners/http-listener/src/test/resources/pentaho-webpackage-2a-package.json
@@ -1,0 +1,4 @@
+{
+  "name": "package-name-2a",
+  "version": "1.0"
+}
