+++ pentaho-webpackage/listeners/http-listener/src/test/resources/pentaho-webpackage-2b-package.json
@@ -1,0 +1,4 @@
+{
+  "name": "package-name-2b",
+  "version": "1.1"
+}
