+++ pentaho-webpackage/listeners/http-listener/src/test/resources/pentaho-webpackage-2c-package.json
@@ -1,0 +1,4 @@
+{
+  "name": "package-name-2c",
+  "version": "1.2"
+}
